feat: 게시글 업로드 시 작성된 이미지들이 S3에서 삭제되지 않도록 cleanup 함수에 조건 추가
