+++ src/components/Trade/Quill/index.tsx
@@ -1,4 +1,4 @@
-import { useRef, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import ReactQuill from 'react-quill';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
@@ -37,11 +37,14 @@ export default function TradeQuill({
   const { modalState, handleModalOpen, handleModalClose } = useModalState();
   const { toastMessageProps, handleToastMessageProps } = useToastMessageType();
   const { state }: { state: { data: TradeBoardDetailType } } = useLocation();
-  const QuillRef = useRef<ReactQuill>();
-  // 이미지를 업로드 하기 위한 함수
-  const modules = useQuillModules(QuillRef, setImages);
 
   const [isProcessing, setIsProcessing] = useState(false);
+
+  const QuillRef = useRef<ReactQuill>();
+  const imagesRef = useRef(images);
+  const isProcessingRef = useRef(isProcessing);
+
+  const modules = useQuillModules(QuillRef, setImages);
 
   const queryClient = useQueryClient();
   const { mutate, isLoading: isUpdateLoading } = useMutation(
@@ -155,6 +158,20 @@ export default function TradeQuill({
   const isUpdating = Boolean(state && !state.data.tmpYn); // 등록된 글을 수정하는 상태
   const isSaving = Boolean(state && state.data.tmpYn); // 임시저장된 글을 작성하는 상태
 
+  useEffect(() => {
+    imagesRef.current = images;
+    isProcessingRef.current = isProcessing;
+  }, [images, isProcessing]);
+
+  useEffect(() => {
+    return () => {
+      if (!isProcessingRef.current) {
+        deleteFile(imagesRef.current);
+        resetImages();
+      }
+    };
+  }, []);
+
   return (
     <div className={styles.container}>
       <section className={styles.sectionWrapper}>
@@ -178,7 +195,6 @@ export default function TradeQuill({
                 QuillRef.current = element;
               }
             }}
-            value={form.code}
             onChange={(value) => setForm((prev) => ({ ...prev, code: value }))}
             modules={modules}
             placeholder="사진 5장 이상은 필수입니다. 5장 이상(건물 외관, 내부 포함) 업로드 되지 않을 시, 반려됩니다."
