+++ src/pages/Trade/Write/index.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState } from 'react';
+import { useRef, useState } from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
 import { FiSearch } from 'react-icons/fi';
 import { MdUploadFile } from 'react-icons/md';
@@ -11,7 +11,7 @@ import {
   TradeBoardDetailType,
   TradeBoardForm,
 } from '@/types/Board/tradeType';
-import { deleteFile, uploadFile } from '@/apis/uploadS3';
+import { uploadFile } from '@/apis/uploadS3';
 import { imageStore } from '@/store/imageStore';
 import userStore from '@/store/userStore';
 import { getRentalPriceType } from '@/utils/utils';
@@ -25,7 +25,7 @@ const { VITE_S3_DOMAIN } = import.meta.env;
 
 export default function TradeWritePage() {
   const { user } = userStore();
-  const { images, setImages, resetImages } = imageStore();
+  const { setImages } = imageStore();
 
   const { state }: { state: { data: TradeBoardDetailType } } = useLocation();
 
@@ -56,7 +56,6 @@ export default function TradeWritePage() {
     state ? state.data.imageUrls[0].split('/')[3] : '',
   );
 
-  const imagesRef = useRef(images);
   const thumbnailRef = useRef<HTMLInputElement>(null);
   // 매물특징
   const [isPostcodeOpen, setIsPostcodeOpen] = useState(false);
@@ -92,17 +91,6 @@ export default function TradeWritePage() {
       }
     }
   };
-
-  useEffect(() => {
-    imagesRef.current = images;
-  }, [images]);
-
-  useEffect(() => {
-    return () => {
-      deleteFile(imagesRef.current);
-      resetImages();
-    };
-  }, []);
 
   if (!user) return <Navigate to="/login" />;
